# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CyberGame API",
        "contact": {}
    },
    "paths": {
...
        "/permissions/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Create Permission",
                "parameters": [
                    {
                        "description": "Create Permission",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreatePermission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/recommends/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Recommend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommends"
                ],
                "summary": "Create Recommend",
                "parameters": [
                    {
                        "description": "Create Recommend",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateRecommend"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recommends/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Recommend List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommends"
                ],
                "summary": "Get Recommend List",
                "parameters": [
                    {
                        "type": "string",
                        "example": "",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessWithList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recommends/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Recommend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommends"
                ],
                "summary": "Update Recommend",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recommend ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Recommend",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateRecommend"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recommends/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Recommend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommends"
                ],
                "summary": "Delete Recommend",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recommend ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scammers/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Scammer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scammers"
                ],
                "summary": "Create Scammer",
                "parameters": [
                    {
                        "description": "Create Scammer",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateScammer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scammers/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Scammer List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scammers"
                ],
                "summary": "Get Scammer List",
                "parameters": [
                    {
                        "type": "string",
                        "name": "bankName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "2021-01-01T00:00:00+07:00",
                        "name": "dateEnd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "2021-01-01T00:00:00+07:00",
                        "name": "dateStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessWithList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "model.CreatePermission": {
            "type": "object",
            "required": [
                "permissions"
            ],
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PermissionName"
                    }
                }
            }
        },
        "model.CreateRecommend": {
            "type": "object",
            "required": [
                "status",
                "title",
                "url"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "maxLength": 10,
                    "enum": [
                        "ACTIVE",
                        "DEACTIVE"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                },
                "url": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "https://www.facebook.com/.../"
                }
            }
        },
        "model.CreateScammer": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "bankAccount": {
                    "type": "string",
                    "maxLength": 15
                },
                "bankname": {
                    "type": "string",
                    "maxLength": 50
                },
                "fullname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 10
                },
                "reason": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "model.PermissionName": {
            "type": "object",
            "required": [
                "name",
                "permissionKey"
            ],
            "properties": {
                "isMain": {
                    "type": "boolean",
                    "default": false
                },
                "name": {
                    "type": "string"
                },
                "permissionKey": {
                    "type": "string"
                }
            }
        },
...
        "model.Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.SuccessWithList": {
            "type": "object",
            "properties": {
                "list": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}